+++ Plutarch/Internal/Subtype.hs
@@ -5,8 +5,6 @@ module Plutarch.Internal.Subtype (
   PSubtype,
   PSubtype',
   pupcast,
-  pupcastF,
-  pdowncastF,
 ) where
 
 import Data.Kind (Constraint)
@@ -15,10 +13,7 @@ import GHC.TypeError (
   ErrorMessage (ShowType, Text, (:<>:)),
   TypeError,
  )
-
 import Plutarch.Internal.PlutusType (
-  PContravariant,
-  PCovariant,
   PlutusType (PInner),
  )
 import Plutarch.Internal.Term (PType, Term, punsafeCoerce)
@@ -66,15 +61,3 @@ type family PSubtype (a :: PType) (b :: PType) :: Constraint where
 
 pupcast :: forall a b s. PSubtype a b => Term s b -> Term s a
 pupcast = let _ = witness (Proxy @(PSubtype a b)) in punsafeCoerce
-
-pupcastF :: forall a b (p :: PType -> PType) s. (PSubtype a b, PCovariant p) => Proxy p -> Term s (p b) -> Term s (p a)
-pupcastF _ =
-  let _ = witness (Proxy @(PSubtype a b))
-      _ = witness (Proxy @(PCovariant p))
-   in punsafeCoerce
-
-pdowncastF :: forall a b (p :: PType -> PType) s. (PSubtype a b, PContravariant p) => Proxy p -> Term s (p a) -> Term s (p b)
-pdowncastF _ =
-  let _ = witness (Proxy @(PSubtype a b))
-      _ = witness (Proxy @(PContravariant p))
-   in punsafeCoerce